refactor(ui): split geodata types into per-concern files
Replaces the 188-line src/types.ts dump with src/types/ split by
concern: geoCommand.ts for classic GEO* command shapes, rqeGeo.ts for
Redis Query Engine types, plugin.ts for the workbench plugin shell. A
barrel index.ts re-exports everything so existing consumers keep
resolving './types' / '../../types' unchanged.
